--- src/estimators/old_code/transferentropy_transferoperator_triang.jl
@@ -1,6 +1,0 @@
-import StateSpaceReconstruction:
-	DelaunayTriangulation
-
-
-function point_representatives(pts, DT::DelaunayTriangulation)
-end
