+++ src/gridestimators/VisitationFrequency.jl
