+++ src/symbolicestimators/SymbolicAmplitudeAware.jl
