+++ src/symbolicestimators/SymbolicPerm.jl
