+++ src/symbolicestimators/symbolic.jl
